refactor(*) remove code redundancy
code and tests are identical. Merged them into a single set of
files, smart enough to detect wether it is pre/post they are running.
